+++ src/main/java/net/frozenblock/wilderwild/worldgen/impl/feature/TubeWormsFeature.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2023-2025 FrozenBlock
+ * This file is part of Wilder Wild.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.frozenblock.wilderwild.worldgen.impl.feature;
+
+import com.mojang.serialization.Codec;
+import net.frozenblock.wilderwild.registry.WWBlocks;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.WorldGenLevel;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.DoublePlantBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.levelgen.Heightmap;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
+import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
+import org.jetbrains.annotations.NotNull;
+
+public class TubeWormsFeature extends Feature<NoneFeatureConfiguration> {
+
+	public TubeWormsFeature(Codec<NoneFeatureConfiguration> codec) {
+		super(codec);
+	}
+
+	@Override
+	public boolean place(@NotNull FeaturePlaceContext<NoneFeatureConfiguration> featurePlaceContext) {
+		WorldGenLevel worldGenLevel = featurePlaceContext.level();
+		BlockPos blockPos = featurePlaceContext.origin();
+		int k = worldGenLevel.getHeight(Heightmap.Types.OCEAN_FLOOR, blockPos.getX(), blockPos.getZ());
+		BlockPos blockPos2 = new BlockPos(blockPos.getX(), k, blockPos.getZ());
+		if (worldGenLevel.getBlockState(blockPos2).is(Blocks.WATER)) {
+			BlockState blockState = WWBlocks.TUBE_WORMS.defaultBlockState();
+			if (blockState.canSurvive(worldGenLevel, blockPos2) && worldGenLevel.getBlockState(blockPos2.above()).is(Blocks.WATER)) {
+				DoublePlantBlock.placeAt(worldGenLevel, blockState, blockPos2, Block.UPDATE_CLIENTS);
+				return true;
+			}
+		}
+		return false;
+	}
+}